refactor(pipeline/network): extract Indexing from NetworkDiscovery
Move response-indexing primitives (parseTextOrNull, shouldRecordResponse,
isUnsupportedUrl, parseResponse, handleResponse, extractBaseUrl,
isReplayablePost, extractRequestMeta, isJsonContentType + WK header
constants and JSON content-type sentinels) out of NetworkDiscovery.ts
into a dedicated Indexing/ sub-module.

NetworkDiscovery.ts continues to re-export the public symbols
(parseTextOrNull / shouldRecordResponse / isUnsupportedUrl /
parseResponse plus their branded signal types and IParsedBody) so
existing test and consumer imports stay byte-identical.

Internal use sites (handleResponse, parseResponse, isReplayablePost,
extractBaseUrl) and the WK header constants (ORIGIN_HEADERS,
REFERER_HEADERS, SITE_ID_HEADERS, BROWSER_STANDARD_HEADERS) are now
imported from Indexing/Indexing.js.

Phase 4 split: commit 3/9. NetworkDiscovery.ts: 1623 → 1319 LoC.
